+++ src/nvim/option.c
@@ -2123,7 +2123,7 @@ void set_init_1(void)
  * Set an option to its default value.
  * This does not take care of side effects!
  */
-static void 
+static void
 set_option_default (
     int opt_idx,
     int opt_flags,                  /* OPT_FREE, OPT_LOCAL and/or OPT_GLOBAL */
@@ -2184,7 +2184,7 @@ set_option_default (
 /*
  * Set all options (except terminal options) to their default value.
  */
-static void 
+static void
 set_options_default (
     int opt_flags                  /* OPT_FREE, OPT_LOCAL and/or OPT_GLOBAL */
 )
@@ -2477,7 +2477,7 @@ void set_title_defaults(void)
  *
  * returns FAIL if an error is detected, OK otherwise
  */
-int 
+int
 do_set (
     char_u *arg,               /* option string (may be written to!) */
     int opt_flags
@@ -3218,7 +3218,7 @@ theend:
  * Call this when an option has been given a new value through a user command.
  * Sets the P_WAS_SET flag and takes care of the P_INSECURE flag.
  */
-static void 
+static void
 did_set_option (
     int opt_idx,
     int opt_flags,              /* possibly with OPT_MODELINE */
@@ -3288,7 +3288,7 @@ static char_u *check_cedit(void)
  * When switching the title or icon off, call mch_restore_title() to get
  * the old value back.
  */
-static void 
+static void
 did_set_title (
     int icon                   /* Did set icon instead of title */
 )
@@ -3309,7 +3309,7 @@ did_set_title (
 /*
  * set_options_bin -  called when 'bin' changes value.
  */
-void 
+void
 set_options_bin (
     int oldval,
     int newval,
@@ -3622,7 +3622,7 @@ static void redraw_titles(void) {
  * When "set_sid" is zero set the scriptID to current_SID.  When "set_sid" is
  * SID_NONE don't set the scriptID.  Otherwise set the scriptID to "set_sid".
  */
-void 
+void
 set_string_option_direct (
     char_u *name,
     int opt_idx,
@@ -3677,7 +3677,7 @@ set_string_option_direct (
 /*
  * Set global value for string option when it's a local option.
  */
-static void 
+static void
 set_string_option_global (
     int opt_idx,                    /* option index */
     char_u **varp             /* pointer to option variable */
@@ -5728,7 +5728,7 @@ static int findoption(char_u *arg)
  *	     hidden String option: -2.
  *		   unknown option: -3.
  */
-int 
+int
 get_option_value (
     char_u *name,
     long *numval,
@@ -5780,7 +5780,7 @@ get_option_value (
 // opt_type). Uses
 //
 // Returned flags:
-//       0 hidden or unknown option, also option that does not have requested 
+//       0 hidden or unknown option, also option that does not have requested
 //         type (see SREQ_* in option_defs.h)
 //  see SOPT_* in option_defs.h for other flags
 //
@@ -6000,7 +6000,7 @@ static int find_key_option(char_u *arg)
  * if 'all' == 1: show all normal options
  * if 'all' == 2: show all terminal options
  */
-static void 
+static void
 showoptions (
     int all,
     int opt_flags                  /* OPT_LOCAL and/or OPT_GLOBAL */
@@ -6070,8 +6070,8 @@ showoptions (
      * display the items
      */
     if (run == 1) {
-      assert(Columns <= LONG_MAX - GAP
-             && Columns + GAP >= LONG_MIN + 3
+        assert(Columns <= INT_MAX - GAP
+                && Columns + GAP >= INT_MIN + 3
              && (Columns + GAP - 3) / INC >= INT_MIN
              && (Columns + GAP - 3) / INC <= INT_MAX);
       cols = (int)((Columns + GAP - 3) / INC);
@@ -6119,7 +6119,7 @@ static int optval_default(vimoption_T *p, char_u *varp)
  * showoneopt: show the value of one option
  * must not be called with a hidden option!
  */
-static void 
+static void
 showoneopt (
     vimoption_T *p,
     int opt_flags                          /* OPT_LOCAL or OPT_GLOBAL */
@@ -7061,7 +7061,7 @@ static int expand_option_idx = -1;
 static char_u expand_option_name[5] = {'t', '_', NUL, NUL, NUL};
 static int expand_option_flags = 0;
 
-void 
+void
 set_context_in_set_cmd (
     expand_T *xp,
     char_u *arg,
@@ -7415,7 +7415,7 @@ void ExpandOldSetting(int *num_file, char_u ***file)
  * Get the value for the numeric or string option *opp in a nice format into
  * NameBuff[].  Must not be called with a hidden option!
  */
-static void 
+static void
 option_value2string (
     vimoption_T *opp,
     int opt_flags                          /* OPT_GLOBAL and/or OPT_LOCAL */
@@ -7472,7 +7472,7 @@ static int wc_use_keyname(char_u *varp, long *wcp)
  *
  * langmap_mapchar[] maps any of 256 chars to an ASCII char used for Vim
  * commands.
- * langmap_mapga.ga_data is a sorted table of langmap_entry_T. 
+ * langmap_mapga.ga_data is a sorted table of langmap_entry_T.
  * This does the same as langmap_mapchar[] for characters >= 256.
  */
 /*
@@ -7812,7 +7812,7 @@ static void fill_breakat_flags(void)
  * Return OK for correct value, FAIL otherwise.
  * Empty is always OK.
  */
-static int 
+static int
 check_opt_strings (
     char_u *val,
     char **values,
@@ -7829,7 +7829,7 @@ check_opt_strings (
  * Return OK for correct value, FAIL otherwise.
  * Empty is always OK.
  */
-static int 
+static int
 opt_strings_flags (
     char_u *val,               /* new value */
     char **values,           /* array of valid string values */
@@ -7911,7 +7911,7 @@ static int check_opt_wim(void)
 /*
  * Check if backspacing over something is allowed.
  */
-int 
+int
 can_bs (
     int what                   /* BS_INDENT, BS_EOL or BS_START */
 )
