+++ src/nvim/option_defs.h
@@ -329,7 +329,7 @@ EXTERN long p_ch;               /* 'cmdheight' */
 EXTERN int p_confirm;           /* 'confirm' */
 EXTERN int p_cp;                /* 'compatible' */
 EXTERN char_u   *p_cot;         /* 'completeopt' */
-EXTERN long p_ph;               /* 'pumheight' */
+EXTERN int p_ph;               /* 'pumheight' */
 EXTERN char_u   *p_cpo;         /* 'cpoptions' */
 EXTERN char_u   *p_csprg;       /* 'cscopeprg' */
 EXTERN int p_csre;              /* 'cscoperelative' */
@@ -471,7 +471,7 @@ EXTERN long p_rdt;              /* 'redrawtime' */
 EXTERN int p_remap;             /* 'remap' */
 EXTERN long p_re;               /* 'regexpengine' */
 EXTERN long p_report;           /* 'report' */
-EXTERN long p_pvh;              /* 'previewheight' */
+EXTERN int p_pvh;              /* 'previewheight' */
 EXTERN int p_ari;               /* 'allowrevins' */
 EXTERN int p_ri;                /* 'revins' */
 EXTERN int p_ru;                /* 'ruler' */
